server/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
transformers>=4.40.0
Pillow>=10.0.0
tzdata>=2024.1
